+++ test/synchronize_test.dart
@@ -2,7 +2,9 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
-@TestOn('vm')
+// Windows sees different bytes than other OSes, possibly because of newline
+// normalization issues.
+@TestOn('vm && !windows')
 
 import 'dart:convert';
 import 'dart:io';
@@ -12,17 +14,20 @@ import 'package:test/test.dart';
 
 import '../tool/grind/synchronize.dart' as synchronize;
 
-void main() {
-  test("synchronized files are up-to-date", () {
-    synchronize.sources.forEach((sourcePath, targetPath) {
-      var source = File(sourcePath).readAsStringSync();
-      var target = File(targetPath).readAsStringSync();
+/// The pattern of a checksum in a generated file.
+final _checksumPattern = RegExp(r"^// Checksum: (.*)$", multiLine: true);
 
-      var hash = sha1.convert(utf8.encode(source));
-      if (!target.contains("Checksum: $hash")) {
-        fail("$targetPath is out-of-date.\n"
-            "Run pub run grinder to update it.");
-      }
+void main() {
+  synchronize.sources.forEach((sourcePath, targetPath) {
+    test("synchronized file $targetPath is up-to-date", () {
+      var target = File(targetPath).readAsStringSync();
+      var actualHash = _checksumPattern.firstMatch(target)[1];
+
+      var source = File(sourcePath).readAsBytesSync();
+      var expectedHash = sha1.convert(source).toString();
+      expect(actualHash, equals(expectedHash),
+          reason: "$targetPath is out-of-date.\n"
+              "Run pub run grinder to update it.");
     });
   });
 }